end_scene_setup
